+++ infrastructure/global/main.tf
@@ -55,9 +55,10 @@ output "lock_table_name" {
 
 output "ecr_repository_urls" {
   value = {
-    api       = aws_ecr_repository.api.repository_url
-    processor = aws_ecr_repository.processor.repository_url
-    indexer   = aws_ecr_repository.indexer.repository_url
+    api         = aws_ecr_repository.api.repository_url
+    processor   = aws_ecr_repository.processor.repository_url
+    indexer     = aws_ecr_repository.indexer.repository_url
+    nixiesearch = aws_ecr_repository.nixiesearch.repository_url
   }
 }